requirements.txt

# ===============================
# CORE SYSTEM MONITORING
# ===============================
psutil

# ===============================
# DATA & ML
# ===============================
numpy
pandas
scikit-learn
joblib

# ===============================
# DASHBOARD / UI
# ===============================
streamlit
streamlit-autorefresh

# ===============================
# NETWORK / IDS
# ===============================
scapy

# ===============================
# VISUALIZATION (IDS TRAINING)
# ===============================
matplotlib
seaborn

# ===============================
# SYSTEM / UTILITIES
# ===============================
pywin32>=307

# ===============================
# OPTIONAL (Tray App – Future)
# ===============================
pystray
Pillow


plyer
pyinstaller
